Split Controllers to be ore readable - ItemController into AusleihController and KaufController - PersonController into AdminController, PersonController - MainController was extracted from both, ItemController and PersonController   and Contains searches and navigation
